+++ client/views/menu-item.js
@@ -3,9 +3,11 @@
 export default function(discovery) {
     discovery.view.define('menu-item', function(el, config, data, context) {
         const { content, onClick } = config;
-        const { text, selected } = data;
+        const { text, selected = false, disabled = false } = data;
 
-        if (typeof onClick === 'function') {
+        if (disabled) {
+            el.classList.add('disabled');
+        } else if (typeof onClick === 'function') {
             el.addEventListener('click', () => onClick(data));
         }